refactor: remove custom event
