+++ src/routes/ErrorPage.tsx
@@ -6,10 +6,6 @@ export default function ErrorPage() {
   const error = useRouteError();
   console.error(error);
 
-  const hardResetHandler = () => {
-    document.dispatchEvent(new Event('search-reset'));
-  };
-
   return (
     <div className="mx-auto my-auto flex flex-col items-center space-y-6 font-pixelify text-lg text-red-500">
       <h1>Oops!</h1>
@@ -17,7 +13,7 @@ export default function ErrorPage() {
       <p>
         <i className="text-red-300">{extractErrorMessage(error)}</i>
       </p>
-      <Link className="button-blue my-4" to="/" onClick={hardResetHandler}>
+      <Link className="button-blue my-4" to="/">
         Home
       </Link>
     </div>
